refactor(auth): relocate decorators into auth directory
