+++ src/auth/decorators/roles.decorator.ts
@@ -1,4 +1,4 @@
 import { Reflector } from '@nestjs/core';
-import { UserRole } from '../users/enums/users.enums';
+import { UserRole } from '../../users/enums/users.enums';
 
 export const Roles = Reflector.createDecorator<UserRole[]>();
